requirements.txt
# Azure Voice Assistant Dependencies

# WebSocket Server
websockets>=12.0

# Azure Speech Services
azure-cognitiveservices-speech>=1.35.0

# OpenAI
openai>=1.0.0

# Environment Variables
python-dotenv>=1.0.0

# Authentication
PyJWT>=2.8.0

# Audio Processing (optional)
numpy>=1.24.0